requirements.txt
boto3==1.9.92
certifi==2018.11.29
jsonschema==2.6.0
python-dotenv==0.10.1
python-telegram-bot==11.1.0
urllib3==1.24.1
